pyproject.toml
[project]
name = "decide-api"
version = "0.1.0"
description = ""
authors = [
    {name = "Your Name",email = "you@example.com"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "anthropic (>=0.69.0,<0.70.0)",
    "sqlmodel (>=0.0.25,<0.0.26)",
    "uvicorn (>=0.37.0,<0.38.0)",
    "fastapi (>=0.118.0,<0.119.0)",
    "pydantic-ai-slim (>=1.0.11,<2.0.0)",
    "pgvector (>=0.4.1,<0.5.0)",
    "openai (>=1.109.1,<2.0.0)",
    "python-dotenv (>=1.1.1,<2.0.0)",
    "pyjwt (>=2.10.1,<3.0.0)"
]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[dependency-groups]
dev = [
    "ipykernel (>=6.30.1,<7.0.0)"
]
